Added "enforce" field to BuildQuery, for being able of enforcing returning lists.
BuildQuery was returning a dictionary of lists if returned rows were
larger than one, and a dictionary of elements if returned rows were
exactly one. In some cases, this causes some trouble when a caller
expects lists, but elements are obtained because there is only one row.

This patch adds a new parameter, "enforce", which by default is None,
and can take as well the value "list". If the value is None, the
previous behavior is respected, for backwards compatibility. If it
is "list", a dictionary of lists is returned always.

In addition, an exception is raised if the sql string received is None,
but only if enforce is "list", again for backwards compatibility.

top_questions_qaforums.py has been touched to use this new parameter,
mostly as a test. All calls that expect a list should use this
parameter, to avoid trouble when resulting row is one...